// packages/@tuist/editor/src/lib/utils/types.ts
export type PathType = string[]

export interface ElementsType {
  [key: string]: ElementType
}

// Used during page preview

export interface ElementSize {
  // Content with padding.
  // content
  c: number
  // top margin
  t: number
  // bottom margin
  b: number
}

export interface LegacySize {
  // Legacy format
  content?: number
  topMargin?: number
  bottomMargin?: number
}

export interface ElementSizes {
  [key: string]: ElementSize
}
// ...
export interface ParagraphOptions {
  // Custom tag
  c?: string
  // Options to display paragraphs
  o?: ElementOptionsType
}

export interface GroupElementRefTypeById {
  [key: string]: GroupElementRefType
}

export interface GroupElementRefType {
  path: PathType
  elem: GroupElementType
}

export type Type = 'P' | 'A' | 'E' | 'T' | 'B' | 'I' | 'B+I'

export interface ElementType extends ParagraphOptions {
  // Type
  t: Type
  // Position
  p: number
  // Inner content (string)
  i?: string
  // Selection
  s?: SelectionType
  // Group
  g?: ElementsType
}
// ...
export interface ChangeType {
  op: 'noop' | 'update' | 'delete' | 'create'
  selected?: boolean
  path: PathType
  pathId: string
  elem: ElementType
  resized?: boolean
}

export interface ElementRefTypeById {
  [key: string]: ElementRefType
}

export interface StringElementRefType {
  path: PathType
  elem: StringElementType
}

export interface ElementNamedType {
  ref: string
  elem: ElementType
}

export interface CompositionType {
  sz?: ElementSizes

  toolbox?: ToolboxOperationValueType
  // Data for custom components.
  data?: { [key: string]: any }
  g: ElementsType
  // Path to object containing selection encoded as 'key.subKey.subsubKey'
  spath?: string
  // selection min paragraph position
  smin?: number
  // selection max paragraph position
  smax?: number
  // Title of the composition (when there is one)
  title?: string
}

// ============== Selections

export interface SelectionPositionType {
  top: number
  left: number
}

export interface CaretSelectionType {
  anchorPath: PathType
  anchorOffset: number
  anchorValue?: string
  position: SelectionPositionType
  stringPath?: string // Only used by setSelection
  type: 'Caret'
}

export interface RangeSelectionType {
  anchorPath: PathType
  anchorOffset: number
  anchorValue?: string
  focusPath: PathType
  focusOffset: number
  position: SelectionPositionType
  stringPath?: string // Only used by setSelection
  type: 'Range'
}

export type SelectionType = CaretSelectionType | RangeSelectionType

export interface ChangesType {
  // Initial data for custom elements (further changes do not go through
  // editor).
  data?: { [elemId: string]: any }
  // Full elements.
  elements: { [elemPath: string]: ChangeType }
  start: ChangeType
  end: ChangeType
  // Special cases where we need to prepare selection/toolbox during op eval.
  selection?: SelectionType
}
// ...
export interface ToolboxOpType {
  type: 'emptyParagraph' | 'paragraph' | 'selection'
  position: SelectionPositionType
}

interface ToolboxNoneType {
  type: 'none'
}

export type ToolboxOperationValueType = ToolboxOpType | ToolboxNoneType